Split test files according to involved libraries
